+++ decidim-demographics/app/helpers/decidim/demographics/application_helper.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Demographics
+    # Custom helpers, scoped to the demographics engine.
+    #
+    module ApplicationHelper
+    end
+  end
+end