+++ decidim-meetings/lib/decidim/meetings/component.rb
@@ -48,6 +48,18 @@ Decidim.register_component(:meetings) do |component|
     exports.serializer Decidim::Meetings::MeetingSerializer
   end
 
+  component.exports :meeting_comments do |exports|
+    exports.collection do |component_instance|
+      Decidim::Comments::Export.comments_for_resource(
+        Decidim::Meetings::Meeting, component_instance
+      )
+    end
+
+    exports.include_in_open_data = true
+
+    exports.serializer Decidim::Comments::CommentSerializer
+  end
+
   component.actions = %w(join)
 
   component.settings(:global) do |settings|
@@ -59,6 +71,7 @@ Decidim.register_component(:meetings) do |component|
     settings.attribute :comments_max_length, type: :integer, required: false
     settings.attribute :registration_code_enabled, type: :boolean, default: true
     settings.attribute :resources_permissions_enabled, type: :boolean, default: true
+    settings.attribute :registration_code_enabled, type: :boolean, default: false
     settings.attribute :enable_pads_creation, type: :boolean, default: false
     settings.attribute :creation_enabled_for_participants, type: :boolean, default: false
   end