+++ decidim-meetings/spec/system/meeting_registrations_spec.rb
@@ -241,7 +241,7 @@ describe "Meeting registrations", type: :system do
           component.update!(settings: { registration_code_enabled: false })
         end
 
-        it "does not show the registration code" do
+        it "shows the registration code" do
           visit_meeting
 
           expect(page).to have_no_css(".registration_code")