+++ decidim-proposals/spec/shared/proposals_wizards_examples.rb
@@ -302,17 +302,5 @@ shared_examples "proposals wizards" do |options|
       end
     end
   end
-
-  # rubocop:disable RSpec/EmptyExampleGroup
-  context "when creating a new proposal" do
-    before do
-      login_as user, scope: :user
-      visit_component
-      click_link "New proposal"
-    end
-
-    it_behaves_like "with address" if options[:with_address]
-    it_behaves_like "without address" unless options[:with_address]
-  end
-  # rubocop:enable RSpec/EmptyExampleGroup
 end
+