+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## Upgrade Notes
 
+- **decidim-core**: Add support for Visual Code Remote Containers and GitHub Codespaces [\6638](https://github.com/decidim/decidim/pull/6638)
+
+### Changed
+
 - **Bump Ruby to v2.7**
 
 We've bumped the minimum Ruby version to 2.7.1, thanks to 2 PRs:
@@ -17,6 +21,7 @@ As per [\#6498](https://github.com/decidim/decidim/pull/6498), the comments comp
 
 ### Added
 
+- **decidim-meetings**: Add functionality to enable/disable registration code [\#6698](https://github.com/decidim/decidim/pull/6698)
 - **decidim-core**: Adding functionality to report users [\#6696](https://github.com/decidim/decidim/pull/6696)
 - **decidim-admin**: Adding possibility of unreporting users [\#6696](https://github.com/decidim/decidim/pull/6696)
 - **decidim-core**: Add support for Visual Code Remote Containers and GitHub Codespaces [\6638](https://github.com/decidim/decidim/pull/6638)